WIP: feat(HeadersHandler helper): Add the 'HeadersHandler' helper to manage all relation with the Headers
