+++ package.json_off
@@ -1,6 +1,6 @@
 {
   "name": "mediafiredl",
-  "version": "0.3.1",
+  "version": "0.4.0",
   "description": "A simple command line software to download mediafire files",
   "main": "index.js",
   "type": "module",
