--- src/helpers/HeadersFileParser/index.ts
@@ -1,1 +1,0 @@
-export { HeadersFileParser } from './HeadersFileParser.ts'
