+++ src/services/Downloader/Downloader.ts
@@ -15,8 +15,7 @@ import { convertMsToTime } from "../../utils/convertMsToTime/index.ts";
 import { checkAndCreateFolder } from "../../utils/checkAndCreateFolder/index.ts";
 import { FileLink, ILinkDetails } from "../FileLink/index.ts";
 import { FolderLink } from "../FolderLink/index.ts";
-import { HeadersFileParser } from "../../helpers/HeadersFileParser/index.ts";
-import { downloadHeaders } from '../../utils/headers/index.ts'
+import { downloadHeaders, IHeaders } from '../../utils/headers/index.ts'
 
 export class Downloader extends Events {
   private concurrencyLimit: number;
@@ -27,7 +26,7 @@ export class Downloader extends Events {
   private downloadedFiles: number;
   private inspect: boolean;
   private beautify: boolean;
-  private customHeaders: Record<string, string> | null;
+  private currentHeaders: IHeaders;
 
   constructor(config: IDownloaderConfig) {
     super();
@@ -39,38 +38,11 @@ export class Downloader extends Events {
     this.linksProcessing = new Set();
     this.invalidLinks = new Set();
     this.downloadedFiles = 0;
-    this.customHeaders = null;
+    this.currentHeaders = downloadHeaders;
   }
 
-  public setCustomHeaders(headersFilePath: string): void {
-    try {
-      const parsedHeaders = HeadersFileParser.parseFile(headersFilePath);
-      const warnings = HeadersFileParser.validateCriticalHeaders(parsedHeaders);
-
-      // Mostrar warnings si los hay
-      if (warnings.length > 0) {
-        console.log(chalk.yellow("\n⚠️  Headers Configuration Warnings:"));
-        warnings.forEach((warning) => console.log(chalk.yellow(`   ${warning}`)));
-        console.log();
-      }
-
-      // Merge con headers por defecto
-      this.customHeaders = HeadersFileParser.mergeWithDefaults(parsedHeaders);
-
-      console.log(
-        chalk.green(
-          `✓ Custom headers loaded from: ${headersFilePath}\n`,
-        ),
-      );
-    } catch (error) {
-      throw new Error(
-        `Failed to load custom headers: ${error instanceof Error ? error.message : error}`,
-      );
-    }
-  }
-
-  private getDownloadHeaders(): Record<string, string> {
-    return this.customHeaders || downloadHeaders;
+  public setCustomHeaders(headers: IHeaders): void {
+    this.currentHeaders = headers;
   }
 
   public addLinks(links: string[], output: string) {
@@ -164,7 +136,7 @@ export class Downloader extends Events {
     output: string,
   ): Promise<void> {
     const response = await fetch(url, {
-      headers: this.getDownloadHeaders(),
+      headers: this.currentHeaders,
     });
     if (!response.ok) {
       throw new Error(
